created proper data modeling type files.
