+++ types/cart.ts
@@ -1,0 +1,6 @@
+export interface CartItem {
+    userId: string;
+    productId: string;
+    quantity: number;
+    addedAt: string; // ISO timestamp
+}
