+++ types/wishlist.ts
@@ -1,0 +1,5 @@
+export interface WishlistItem {
+    userId: string;
+    productId: string;
+    addedAt: string; // ISO timestamp
+}
