+++ utils/db.ts
@@ -38,196 +38,6 @@ export async function collectValues<T>(iter: Deno.KvListIterator<T>) {
   return await Array.fromAsync(iter, ({ value }) => value);
 }
 
-// identity, authentication, and permissions
-
-type UserRole =  'customer' | 'brand' | 'admin'  | 'super';
-
-export interface User {
-  id: string; // ULID or UUID
-  login: string;
-  email: string;
-  passwordHash: string; // tbd hash encryption
-  role: UserRole;
-  createdAt: string; // ISO timestamp
-  lastLoginAt?: string; // for auditing / UX
-  isDisabled?: boolean; // soft bans, inactive accounts
-}
-
-type ProfileStatus = 'active' | 'pending' | 'inactive';
-
-// public-facing display layer
-interface BaseProfile {
-  id: string;
-  userId: string;
-  createdAt: string;
-  status: ProfileStatus;
-}
-
-interface BrandProfile extends BaseProfile {
-  title: string;
-  slug: string;
-  bio?: string;
-  website?: string;
-  logoUrl?: string;
-}
-
-interface CustomerProfile extends BaseProfile {
-  displayName?: string;
-  avatarUrl?: string;
-  preferences?: Record<string, unknown>;
-  language?: string;
-}
-
-// product-related
-export interface Product {
-  // Uses ULID
-  id: string;
-  brandId: string;
-  title: string;
-  price: number; // price in BRL
-  description: string;
-  mainImage: string;
-  images: ProductImage[];
-  categoryId: string; // Having a single source of truth (ProductCategory) for translations means consistent i18n.
-  tagIds: string[]; // "Grunge", "Cyberpunk", "Genderless", "Made in Brazil"
-  createdAt: string; // ISO timestamp
-  inventory?: number | null; // null = infinite
-  isAvailable?: boolean; // to toggle visibility
-  updatedAt?: string; // for sorting / change detection
-}
-
-export interface ProductCategory {
-  id: string; // e.g., "tops", "outerwear"
-  name: Record<string, string>; // Localized names
-  icon?: string; // Optional icon/image for UI use
-  order?: number; // For sorting categories in UI
-}
-
-export interface ProductTag {
-  id: string; // e.g., "grunge", "genderless"
-  name: Record<string, string>; // Localized tag names
-  color?: string; // Optional for badges
-  description?: Record<string, string>; // If tags need context
-}
-
-interface ProductImage {
-  url: string;
-  alt?: string;
-  priority?: number;
-}
-
-export interface CartItem {
-  userId: string;
-  productId: string;
-  quantity: number;
-  addedAt: string; // ISO timestamp
-}
-
-export interface WishlistItem {
-  userId: string;
-  productId: string;
-  addedAt: string; // ISO timestamp
-}
-
-// Order
-
-type OrderStatus = 'pending' | 'paid' | 'shipped' | 'cancelled' | 'delivered';
-type OrderItem = { productId: string; quantity: number };
-
-export interface Order {
-  id: string;
-  userId: string;
-  items: OrderItem[];
-  total: number;
-  placedAt: string;
-  notes?: string;
-  status: OrderStatus;
-  shippingAddressId: string;
-  paymentMethodId: string;
-  couponId?: string;
-  discountAmount?: number;
-}
-
-interface Coupon {
-  id: string;
-  code: string;
-  description?: string;
-  discountType: 'percentage' | 'fixed';
-  value: number; // e.g., 20 for 20% or R$20
-  expiresAt?: string;
-  maxUses?: number;
-  userLimit?: number;
-}
-
-type ReturnStatus = 'requested' | 'approved' | 'rejected' | 'refunded';
-
-interface ReturnRequest {
-  id: string;
-  orderId: string;
-  userId: string;
-  items: OrderItem[];
-  reason: string;
-  status: ReturnStatus;
-  createdAt: string;
-  resolvedAt?: string;
-}
-
-export interface Address {
-  id: string; // ULID or UUID
-  userId: string;
-  name: string;
-  street: string;
-  city: string;
-  state: string;
-  zip: string;
-  country: string;
-  phone?: string;
-  isDefault: boolean;
-}
-
-type ShippingCarrier = 'correios' | 'custom' | 'manual';
-type ShippingStatus = 'pending' | 'in_transit' | 'delivered' | 'failed';
-
-interface Shipping {
-  id: string;
-  orderId: string;
-  carrier?: ShippingCarrier;
-  trackingCode?: string;
-  estimatedDelivery?: string;
-  status: ShippingStatus;
-  lastUpdated: string;
-  history?: ShippingHistoryItem[];
-}
-
-interface ShippingHistoryItem {
-  status: ShippingStatus;
-  timestamp: string;
-  location?: string;
-  note?: string;
-}
-
-type PaymentMethodProvider = 'stripe' | 'pix';
-
-export interface PaymentMethod {
-  id: string; // reference to Stripe setup ID or similar
-  userId: string;
-  provider: PaymentMethodProvider;
-  label: string; // e.g., "Visa ending in 4242"
-  lastUsedAt?: string;
-  isDefault: boolean;
-  metadata?: Record<string, unknown>; // for stripe/pix token stuff
-}
-
-interface AuditLog {
-  id: string;
-  userId?: string;
-  action: string; // "update_product", "delete_order"
-  targetId?: string;
-  targetType?: string; // "Product", "Order", etc.
-  timestamp: string;
-  metadata?: Record<string, unknown>;
-}
-
 /**
  * Random data generators for testing & seeding.
  *
